+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 
 I will try to keep track of changes between releases here
-
+- If /iCloud is an iOS mount, it will be synced to ensure normal operation
 - New rewrite of handling hostname for iOS >= 17 now can use /etc/hostname in traditional way
 - added hostname_ settings to tools/utils.sh related to alt hostname handling iOS >= 17
 - test imgs using TZ=UTC